requirements.txt
streamlit
langchain
openai
faiss-cpu
pytesseract
pillow
PyMuPDF